requirements.txt
streamlit
pywebview

mss
Pillow
pywin32;
onnxocr
